fixup zebra table markup after commit e2b4528
The recent change to renderTableViewer() had a side effect in that the
newly added THEAD element numbered its TRs from 1 and the subsequent
TBODY element numbered its TRs from 1 (once again) making the resulting
table markup incorrect at the top.

Solve this by specifying the "zebra" and "zebra0" classes to work within
TBODY only. This way all the existing code that keeps all TRs in the
implicit TBODY continues to work as before and the code that puts any
header TRs into a THEAD can just use the "zebra0" class for the table.
Update renderTableViewer() to do exactly that and be slightly simpler.